test_gesture development (finished but still working out in distinguishing cases when length > 0 and < 0 to make program work properly in both cases)

+++ lab4/lab4.c
@@ -73,7 +73,7 @@ static int proc_args(int argc, char **argv)
 		length = parse_ulong(argv[2], 10);
 		if (length == ULONG_MAX)
 			return 1;
-		printf("mouse::test_gesture(%lu)\n", length);
+		printf("mouse::test_gesture(%d)\n", (short)length);
 		return test_gesture(length);
 	}
 	else {
